third_party/blink/web_tests/fast/js/mozilla/strict/15.3.5.1-expected.txt

PASS 'use strict'; var f = fn(); f.length = 1; f.length threw exception of type TypeError.
PASS var f = fn(); f.length = 1; f.length is 3
PASS true === true
PASS 'use strict'; var f = fn(); delete f.length is true
PASS var f = fn(); delete f.length is true
PASS true === true
 PASSED! 
PASS successfullyParsed is true

TEST COMPLETE
